remove unused utils
